requirements.txt
certifi==2018.4.16
chardet==3.0.4
defusedxml==0.5.0
Django==1.11
django-contrib-comments==1.8.0
django-oauth-plus==2.2.9
django-oauth2-provider==0.2.6.1
django-reset==0.2.0
djangorestframework==3.6.2
djangorestframework-oauth==1.1.0
httplib2==0.11.3
idna==2.6
oauth2==1.9.0.post1
oauthlib==2.0.7
PyJWT==1.6.1
python3-openid==3.1.0
pytz==2018.4
requests==2.18.4
requests-oauthlib==0.8.0
shortuuid==0.5.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.22
uWSGI==2.0.15